+++ components2/src/components/s-ds-header/s-ds-header.tsx
@@ -1,4 +1,4 @@
-import { Component, Host, h } from '@stencil/core';
+import { Component, Host, h, Prop } from '@stencil/core';
 
 @Component({
   tag: 's-ds-header',
@@ -6,11 +6,12 @@ import { Component, Host, h } from '@stencil/core';
   shadow: true,
 })
 export class SDsHeader {
+  @Prop() headerTitle = 'Default header text';
 
   render() {
     return (
       <Host>
-        <header>header</header>
+        <header>Scania {this.headerTitle}</header>
         <slot></slot>
       </Host>
     );