+++ site/src/App.js
@@ -4,11 +4,10 @@ import './App.css';
 function App() {
   return (
     <div className="App">
-      <s-ds-header></s-ds-header>
       <my-header></my-header>
-      <my-footer></my-footer>
-      <s-ds-footer></s-ds-footer>
+      <main>test</main>
       <my-cookie something-else="dsadas name"> <span>test</span> </my-cookie>
+      <my-footer></my-footer>
     </div>
   );
 }